refactor MyHandler.ts
