+++ src/classes/MyHandler/export-process.ts
@@ -1,0 +1,4 @@
+import processAccepted from './offer/accepted/process';
+import updateListings from './offer/accepted/updateListings';
+
+export { processAccepted, updateListings };
