+++ src/classes/MyHandler/offer/notify/accepted.ts
@@ -1,0 +1,11 @@
+import { TradeOffer } from 'steam-tradeoffer-manager';
+import Bot from '../../../Bot';
+
+export default function accepted(offer: TradeOffer, bot: Bot): void {
+    bot.sendMessage(
+        offer.partner,
+        bot.options.customMessage.success
+            ? bot.options.customMessage.success
+            : '/pre ✅ Success! The offer went through successfully.'
+    );
+}
