+++ src/classes/MyHandler/offer/review/export-review.ts
@@ -1,0 +1,3 @@
+import sendReview from './send-review';
+
+export { sendReview };
