+++ src/lib/DiscordWebhook/sendAlert.ts
@@ -3,7 +3,7 @@ import { XMLHttpRequest } from 'xmlhttprequest-ts';
 import timeNow from '../tools/time';
 
 import Bot from '../../classes/Bot';
-import MyHandler from '../../classes/MyHandler';
+import MyHandler from '../../classes/MyHandler/MyHandler';
 
 export default function sendAlert(
     type: string,
